Bug fix for output of residual vector in parallel when using STK.
Previously, the residual vector was copied into the STK nodal fields
from a non-overlapping vector.  This resulted in the STK data for ghosted
nodes not having the proper value of the residual.  The situation propagated
to the exodus output files, where one of the parallel files would have
the correct value and the others (if they had ghosted copies of the node)
would have a value of zero.  Post-processing tools like epu would then do
something arbitrary to decide which value was the correct value when combining
parallel output files.

The copying of the residual vector to STK containers is and always has been
a one-off function.  This call is carefully placed in Albany_Application.cpp
after the internal force has been computed but before the residual vector
is modified for boundary conditions.  In this way, one can (perhaps not
safely) assume that the residual vector includes reaction forces for mechanics
problems.

To be clear, the bug affected output only and didn't cause problems with
the Albany solution procedure.