Updated interface to LAME material models.
This commit brings the Albany interface to LAME up to date with
the development version of LAME.  There were several interface
changes and changes to the LAME directory structure.  Also, LAME
now depends on the Sierra math toolkit.  Thus linking Albany to
LAME requires that one has a compatible version of liblame.a and
libmath_toolkit.a.  To the best of my knowledge, there is no
current mechanism to build a standalone math toolkit, I'll work
on that.  For now, I am able to build and test the Albany LAME
interface by pointing cmake to the corresponding include and lib
directories within a standard Sierra build (bake).