config: Make Albany always inherit Trilinos C++ flags.
Previously, if CMAKE_CXX_FLAGS was set (or, as Gary T. pointed out, if CXXFLAGS
was an env variable), Albany would not inherit Trilinos's C++ flags. The most
obvious problem is -std=c++11 might then be missed. More subtly, if -DNEBUG is
used in the Trilinos build but not the Albany one, there is an stk_mesh link
issue.

A solution I discussed with Gary and Andy is to always force inheriting the
flags. There are use cases where one might not want this (e.g., a -ggdb build of
Albany against an -O3 build of Trilinos, or different warning on/off settings),
but on the whole, the pros (most importantly, more likely to avoid
out-of-the-box build problems) outway the cons.

Thanks to Gary and Andy for the discussions.