LCM: Remove lapack object. The rationale for this new minisolver is to replace the dependency on Lapack with one on minitensor so that it can be Kokkoized.
The minisolver uses a brute-force approach by calling the inverse
algorithm in minitensor which is very accurate but not very fast.
It is believed that more benefit will be obtained by Kokkoizing
the constituive models in this fashion because the systems to
solve are very small but there are many, many of them.

If profiling indicates that the solve method is unacceptably slow,
more sophisticated algorithms can the used as drop-in replacements,
but this only after hard data is collected (premature optimization
is the root of all evil [Knuth]).