Improve performance of MPJacobian scatter evaluator.
The scatter evaluator for the MPJacobian evaluation type was summing in
one value at a time, which is much slower than a whole row at a time due
to the search over column IDs that must take place.  I changed the
scatter to pull out a whole matrix row of values for each ensemble
value, and then sum in the whole row.  This substantially improved the
performance of the MPJacobian scatter.