Add --vtune command-line argument for performance profiling.
I added "--vtune" as a command-line argument for all Albany drivers
which connects the application to VTune for performance profiling.  It
does this by calling the VTune command-line collector (amplxe-cl) which
stores its results in a local subdirectory called "vtune".  For this to
work you just need to have the amplxe-cl command in your path.  It also
works for MPI parallel runs by creating a separate VTune result for each
MPI rank.  I thought it made more sense to add this as a command-line
argument instead of a parameter in the input file so that you don't need
to modify the input file to run it.

To facilitate this, I also created a little command-line arguments
struct in Albany_Utils.hpp to store and parse command-line arguments.  I
changed each driver to use it.  I believe I left the existing default
filenames intact (which varied depending on the driver).