Kokkos: Move resumeFill/fillComplete out of evaluator.
ALBANY_KOKKOS_UNDER_DEVELOPMENT was hanging on some MPI test problems. That's
because if the number of worksets isn't equal across processes, calling
fillComplete and resumeFill in an evaluator is an MPI error: not all processes
join in the synchronization events. I've moved these calls to the top level.

However, I'm not really sure fillComplete'ing so that getLocalMatrix is valid is
the intended usage. We should consult with the Tpetra and Kokkos devs to find
out how they envision a Jacobian fill to be implemented.

But at least for the moment, get a number of the tests working.