Removed the "required fields" entry in the input parameter sublist "Problem" in the StokesFO examples. Instead, we default it to the usual list of fields names.
Furthermore, in the mesh struct, we read the required fields info only if present,
otherwise the code was defaulting the field type (vector/scalar and node/elem)
and generated errors when the corresponding info in the input parameter list
was not found. Now we only process required fields info if the list is present.