Add additional nvcc_wrapper script.
Make a new compiler wrapper script "nvcc_wrapper_deluxe" that
allows one to specifically enumerate files and Trilinos packages
that should be compiled with the host compiler (not nvcc).

Note that this doesn't always let one avoid nvcc compilation errors - if the
error lives in a header or definition file, all files that include that must
be listed which may not be practical.