GOAL: adding a separate mechanics problem
This mechanics problem is a stripped down version
of the mechanics problem in LCM, reusing all of the
existing constitutive model machinery developed by
LCM. This problem currently only supports solving for
kinematic degrees of freedom.

The justification for introducing this problem is so
that evaluators and field managers that are specific
to solving an auxiliary adjoint problem can be set up
directly in the problem itself. By introducing this
problem in GOAL, all of this extra baggage is nicely
sandboxed and prevents large and potentially undesirable
changes to the current LCM mechanics problem.