Aeras: adding function that applies laplace*mass^(-1)*laplace in Aeras::HVDecorator.  This will be called in the evalModelImpl, where I put a call to this function, and some notes afterwards.
Note that when calling this function on a singular mass matrix, it will
produce nans (b/c 1/0 = nan).  Function needs to be tested on an example
with a non-singular mass, when a "HyperViscosity2" or similar method is added
as a ParameterList option.

We may want to throw an exception if a 0 in the mass matrix's main diagonal is
detected.