Aeras: temporary fix to avoid calling writeSolutionToMeshDataBaseT and writeSolutionToFileT for line spectral elements (for xz-hydrostatic).  For some reason, the observer and writeSolutionT are called even when no exodus output file is requested in the input file, and the code crashes for line spectral elements. Need to look into why.
Aeras spectral linear elements produce the same solution as isoparametric linear elements on XZHydrostatic
example in serial as well as parallel.  One issue I noticed not including the observer is the total volume response
is not the same for the spectral linera elements as the isoparametric ones.  Does this response work?  Is there something in
it hard-coded for isoparametric elements?

I have not tested Aeras spectral higher order elements yet on xz-hydrostatic, but that is next.

Added "Debug Output" section to input files to dump final solution to matrix market file to do comparisons b/w aeras spectral
and isoparametric elements for XZHydrostatic problems.