PUMI: Slight tweak to clever new XML end logic.
On initial write, use ios::out only and not ios::in | ios::out. An alternative is
   in | out | trunc,
but I see no advantage to this. When the pvd is first created, it should be
created from scratch.

This fixes the following two situations: the pvd exists already and some funny
write patterns happen as it the file is modified, or it does not exist and
cannot be created because of using ios::in without ios::trunc.