LCM: Changed residuals for both the bulk and surface elements
I am checking in this source for Guy Bergel. We found a bug in that the
diffusion residual is normalized by the diffusivity. That's fine if the entire
body has the same diffusivity. However, we are exploring fast pathways for
diffusion in which some blocks contain elevated diffusivities. In addition,
we are also exploring surface elements having increased diffusivities.

Given that the diffusivity is spatially variable, one cannot normalize by
the diffusivity in the Laplacian term. This checkin removes that issue.

The results for a homogeneous problem will not change. I do, however, need
to update the gold files because the residual at the essential boundary
condition is now different (unscaled). The concentrations are unchanged.