LCM: Final bug fix for stabilized pressure, I'm thinking.
This one was fun. I noticed that the pressure-pressure part of the Jacobian was
rank deficient by one. The math says otherwise, so the hunt was on.

Turns out FirstPD was writing to stress_, which is a dependent, not evaluated,
field. This was messing up the calculation. Fixing it required rearranging
FirstPK considerably. This fixes the same bug for pore pressure. That doesn't
mean pore pressure necessary had a problem; invalid writing to stress_ matters
only if the evaluation DAG puts another evaluator that depends on Cauchy_Stress
downstream.