Progress on explicit dynamics with Velocity Verlet for Peridigm-Albany coupling.
The WaveInBarPD test is a reproduction of a Peridigm regression test.  The Albany
version matches the Peridigm version, which is nice.

Density and cell volume are currently hard-coded for the meshfree peridynamics
mass matrix.  Obviously needs to be generalized.

The WaveInBarFEM test is a wave passing through a standand FEM mesh (no peridynamics).
The results are encouraging, but it's not clear how the density is being handled.