Build System: Added  "ALBANY_LIBRARIES_ONLY" config option
When turned on with  "-D ALBANY_LIBRARIES_ONLY=ON", the code
will not link any executables and enable no tests. Should be
combined with ENABLE_INSTALL and CMAKE_INSTALL_PREFIX.
This is a request of ice sheet collaborators who use
our libraries but don't want the expense of linking
executables on LCFs file systems.

This addresses bug # "Matt sent me an email today"