LCM: change behavior of linear elastic model when temperature is present
Ottmar Klaas indentified an issue with the thermal expansion term in linear
elasticity. Interested parties can refer to eq. 6.2.29 in Malvern.

The thermal contribution to the stress was

\sigma^{th} = \beta \Delta T \delta_{ij}

And \beta was the "Thermal Expansion Coefficient" input. This caused some
confusion. What is now in the code is

\sigma^{th} = 3 \kappa \alpha \Delta T \delta_{ij}

now \alpha is the "Thermal Expansion Coefficient", and this gives expected
results, without having to scale the input by the bulk modulus.