Aeras: checking in input files for higher order (>2 nodes/element) + xz hydrostatic equations. There is an input file to run with RK4 and Backward Euler.  Comments:
- With node numbering bug fix earlier today, higher order spectral elements run.  Jacobians are no longer singular.
- With backward Euler, both problems converge to a solution with a reasonable average value.  With RK4, in contrast,
the solution blows up even with reduced time step (dt = 0.025) when running up to longer than T = 0.1.  Is there
something about the equations that will make solving as is with an explicit scheme difficult?
- Periodic BCs for higher order spectral elements appear to be OK.
- I believe the Aeras::TotalVolume is not using the right weights for higher order spectral elements.  Need
to figure out what weights to use intead of weighted_measure.  weighted_measure sums to a different value depending on the
element order: L for line elements with 2 nodes, L/2 for line elements with 3 nodes, etc. where L is the length of the element.