Inverted the order of the CXX flags from Albany and Trilinos.
This is to allow debug builds of Albany linking against opt builds of trilinos,
which are useful to reduce the linking time and size (which otherwise make the
debug build MUCH slower than the opt one).

Building against opt builds of trilinos is a good idea if we don't need to
open up and inspect Trilinos data structures during debug, since it makes
the compilation/execution faster, without compromising the debugging.