GOAL: initial attempt at higher order fields
trying to get the goal discretization to understand
higher order fields. the goal discretization is
derived from the pumi discretzation (which is derived
from an apf discretization) to hopefully make everything
convenient during mesh adaptation. because of the 3 levels
of inheritance, these 3 classes needed to be reworked a
bit so that the virtual methods defined at the lowest
inheritance level are actually used.