LCM: Changes for validation and usage: hide some members and expose only what is necessary for configuring solvers and associated methods. Introduce static_assert to prevent invalid instantiation.
The static_assert used in the Funcion_Base struct to prevent
instantiation with FAD types had to be reverted. This is because the
crystal plasticity model uses a derived Function class that is
instantiated with ScalarT that may be a FAD type.

This will need to be changed because Funcion classes are to be
instantiated with non-FAD types. On the other hand, the value(),
gradient() and hessian() member functions are templates themselves
and they admit any evaluation type including FAD types.

If the Function class is instantiated with as FAD type there will be
unnecessary computations for sure. In addition, if the Function class
type and the value/gradient/hessian member type happen to be the same
FAD, the Sacado machinery will operate on both of them without distinction
leading to wrong results.