Add configure flag USE_PUBLICTRILINOS and use it for the MueLu config code.
When the internal and public Trilinos repos diverge, public repo users can use
    -DUSE_PUBLICTRILINOS=ON
which defines
    ALBANY_USE_PUBLICTRILINOS
Then we can special-case code that is breaking because of the desync. Once a
desync is over, we'll remove the special case. This of course requires manual
intervention each time; this flag is just meant to provide the mechanism to do
this manual intervention cleanly.

I've used this mechanism to handle the MueLu config code.