Successfully hooked up MiniSolver within the CrystalPlasticity model.
Plumbing is in place to call MiniSolver to solve the nonlinear system
for implicit integration within the CrystalPlasticity model.  The
MiniSolver appears to be getting the same solution as the home-rolled
solver.

The MiniSolver is not currently able to return the sensitivity information
that is required by the CrystalPlasticity model.  It's currently hooked
up in a test mode in which it solves the nonliner system after the
home-rolled solver has already solved it (and acquired the required
sensitivities for Tangent/Jacobian calls to the constitutive model).
The MiniSolver results and the home-rolled solver results are printed
to std::out for comparison.

I'll keep working with Alejandro (co-design!) as he brings the MiniSolver
to the point where it can produce the data that is needed for patching
into AD types.

If anyone wants to run non-trivial crystal plasticity simulations
with the implicit integrator using this version of the code, they
should undefine MINISOLVER at the top of CrystalPlasticityModel_Def.hpp.

 LocalWords:  src