Aeras: Hackasaurus rex is hungry for speedup.
In the case that there is only one workset, the MDFields containing the basis
function data contain valid data once initialized.

Modify the existing memoization hack to check for and capitalize on this
case. If it holds, then the ComputeBasisFunction eval time goes to essentially
0.

This is a hack for two reasons:

1. The obvious one.

2. The workset size information is not available when the state fields are set
up. So memory in the state manager is being used needlessly if #worksets is 1.

(One mitigating factor is that the hack is localized entirely to the evaluator
itself.)

I want to push this so we can start running with (and checking for) essentially
no basis function cost. Ideally, we'll think of something just as convenient and
fast but not as hacky.