Changes in AbstractDiscretizations concerning side discretizations stuff:
Changed a typedeffed name and changed the return type of two getters: now we
return std::map's rather than Teuchos::RCP to them. After all, one can just
check the size of the map to see if side discretizations are present (instead
of checking the pointer). The idea is to remove one box layer: these structures
do not need to be encapsulated in a pointer; an std::map is already enough.