Added an evaluator for the basal residual.
The basal friction condition residual is now evaluated with the "volume" field
manager, rather than being inside the neumann field manager. This has one big
advantage: it allows to use information from other volume field managers in the
computation. In particular, we can use the value of the velocity norm itself.

StokesFOResid now has an additional input field, called 'basalResid'. However,
this field is required only in case of "sliding". This means that, in case of
full dirichlet BC at the ice-bedrock interface, no evaluator is needed. This is
because the sliding condition requires an additional section in the xml input
file. In order to keep backward compatibility and avoid to modify many xml files
for previous examples (which would not use the new evaluators anyways), the
input field basalResid is required only if the sublist 'FELIX Basal Friction Coefficient'
is present (in which case, also the Basal Friction Coefficient evaluator will
be created.