Addition: side set equations.
We accept that some equations may be defined only on some side set (such as for coupled 3D-2D problems).
This require careful handling of the Jacobian matrix and the maps. In particular,
for equations defined only on one (or more) sideset, the Jacobian is diagonal for the rows
corresponding to nodes NOT on the given sideset(s). Then, the entries of the Jacobian/Residual
in these rows must be properly filled, for instance with 1 for Jacobian and 0 for residual.

Here, we also provide a new Dirichlet BC to handle the jacobian/residual rows corresponding to
"off-node-set" nodes. That is, this BC handles the rows that are not on the selected nodeset
and puts J=1 and res=x-value.

Note: the nodeset corresponding to a given sideset is now created in STKDiscretization. In particular,
after the creation of sidesets, a nodeset is created for each sideset, to contain all the owned nodes
on said sideset.
WARNING: if a nodeset with the same name as the sideset already exists, STKDiscretization ASSUMES
it already contains all the nodes on the sideset. This may not be true! For safety, never give a nodeset
the same name of a sideset, so that STKDiscretization will have those names free for creating the nodesets
corresponding to the sidesets.