Establish an ordering for DBC evaluators.
Phalanx was updated yesterday. The evaluator DAG order, when multiple paths are
possible, has changed. In principle this reordering should not change
answers. However, in practice there is at least one situation in which it does
that we want to keep (based on a discussion with Andy): the case of DBCs that
meet at a corner with incompatible values.

To keep this case working, force an ordering of DBC evaluators by making a chain
of dummy dependencies. The chain orders the DBCs in the order of the
ParameterList iterator. If the PL is populated from an XML file, that is text
order. It turns out that text order is the old order of evaluation, which means
we don't have to rebaseline tests that have this corner case.