LCM: Handle some pointer stuff.
FO_GIS_GisCoupledThickness was broken. Turned out to be because of an
uninitialized variable that was created the other day in Application.

However, I suspect that variable was created based on a red herring. Long story
short: the NotchedCylinder problem still had memory issues, but now only
intermittently (depending on system) visible.

The real issue is that SchwarzMultiscale creates a temporary SolverFactory, then
calls SolverFactory::getParameters() and creates a weak RCP to the object. There
are two problems: (1) The RCP needs to be strong because otherwise the
ParameterList gets deleted when the temporary SolverFactory goes out of
scope. (2) getParameters returns a ref, not an RCP, so one doesn't even have a
handle to the memory in the first place. Solution: Copy the ParameterList ref
from the temp SolverFactory (the copy is necessary since SchwarzMultiscale
doesn't have access to the RCP SolverFactory has) into a strong RCP.

An alternative to the PL copy ctor is to keep SolverFactory as a class array.

A subtler issue involves ownership of the PL. That could be solved multiple
ways. I decided the cleanest thing is to make SchwarzMultiscale take ownership
of it (since it essentially creates it in the first place and has exactly the
right persistence) and give everyone else weak RCPs. The same thing has to
happen if someone decides to keep a SolverFactory array instead of doing what I
did.

I cleaned some surrounding code up while I was at it.