GOAL: fill in computeNodeSets for discretization
Higher order modes for hierarchical shape functions
are not interpolatory. BC behavior for these nodes
needs to be modified to reflect this fact. This commit
is the first step for applying such boundary conditions.
Node sets for the GOAL discretization are constructed
with knowledge of the node type (higher order or not).