Refactoring of the FELIX Hydrology classes
The hydrology problem is now available in two flavors:

- steady: one nonlinear elliptic equation for the effective pressure, with
          drainage sheet depth assumed known
- unsteady: one nonlinear elliptic equation for the effective pressure and
            one evolution equation for the drainage sheet depth, coupled together

Moreover, the BasalFrictionCoefficient evaluator needs to be able to work for
both the hydrology and stokes. Since for Stokes it evaluates a side field and
for the hydrology a cell field, a distinction must be made to properly set the
input/output fields.