Some bugfix in side set mesh handling and field loading.
For the side meshes, we have to be careful to the order of operations:
- entities extraction
- field declaration
- field loading

The first thing to do is field declaration, then entities extraction and
finally field loading. There was a problem with this, when dealing with
a side mesh (other than the basal side) for extruded meshes, since the
order was not handled correctly, although it worked for other meshes.
Therefore, we opted for a rearrangement of the side set mesh methods, and
how they are called from inside GenericSTKMeshStruct.

There is also a bug fix in GenericSTKMeshStruct, related to the method
loadRequiredFields. In particular, when the mesh is a side mesh, there is
the chance that the nodes/elems GIDs do not start from 0, nor they are
contiguous. Therefore, the serial map used to load data from file cannot
be constructed linearly from 0 to numGlobalElements, but instead has to
be created using the Tpetra utility computeGatherMap(...), which creates
a map on the root process, with all the GIDs of the input map (across all
processors).