Minor change in SaveSideSetStateField
Removed the "trick" to add 10 for vector quantities. This is absolutely
useless, since the code works fine for any dimension of the input field/state
pair. There's no need to check whether or not the field is scalar, vector
or even tensor (although, not yet implemented here), since the code simply
switches on the dimension of the data layout, which is enough.