Bug fix in FELIX::ResponseSurfaceVelocityMismatch
The w_measure field was used for both basal and surface sides. This was
wrong, since surface fields are defined on a single side set. Hence, we
have to use two different fields for the measure on the basal and surface
sides. To be clear, we modified also the inv_metric field to be called
inv_metric_surface. Although it is only used in the surface calculation,
this name emphasizes the fact that, even if needed due to some future
addition, this inv metric field should not be used in the basal side
calculations.