Fixed bug in FELIX
In BasalFrictionCoefficientGradient, we moved the Teuchos check on the type
of beta inside the evaluation method. In fact, in StokesFO, the evaluator
is always created, but it is only used in case the user asks for the
response ResponseSurfaceVelocityMismatch. Hence, the test in the constructor
would make simulations that do not use given beta fail even if they do not
use that response.

Furthermore, in the ResponseSurfaceVelocityMismatch, we add gradBeta as a
dependent field only if the regularization parameter is different from 0.