Added evaluator Nodes->Cell interpolation
It is similar to QuadPoints->Cell interpolation, and one could get away
without this evaluator, simply using the combination DOFInterpolation and
QuadPointsToCellInterpolation. However, I created this evaluator to "skip"
one step and to the interpolation in a single evaluation, given that the
body of the evaluator is pretty small anyways.