LCM: add test for equilibrium concentration BC
The BC supports the H/He diffusion and decay effort in that the Dirichlet BC
for the concentration can be specified as a function of the pressure. The
current test is a simple 2D problem that gives the correct value for
concentration on the proper nodeset at the current pressure.