Hooked up MiniSolver with the CrystalPlasticity model.
The code is awash with #defines.  The HOMEROLLEDSOLVER #define,
which is currently turned on, activates the old non-MiniSolver
nonlinear solver.  The MINISOLVER #define, currently turned off,
activates the MiniSolver version of things.  If they're both
turned on, they will both run but the MINISOLVER solution will
be used.  An additional #define, PRINTSOLVERDATA, prints data to
stdout to allow for head-to-head comparison of the two solvers.

If HOMEROLLEDSOLVER is turned off and MINISOLVER is turned on,
the code will compile and the tests will run, but several of them
diff.  More work is required to figure out why this is.  It may be
just a matter of the tolerances for the nonlinear solvers.

The CrystalPlasticityNLS class now uses the peel function that was
recently added by Alejandro.  It appears to be working correctly.
Thanks, Alejandro.

There was a issue with derivative information being lost due to
a filter on slip-system activity that clipped the slip increment
to zero when it was very small.  The filter is now wrapped in
a #ifndef MINISOLVER, such that it is turned off when the MiniSolver
is turned on.  Without this, MiniSolver will throw an exception
because it expects derivative information and will get none.

The path forward is to get the MiniSolver working to the point
where we can rip out the home-rolled version.  This will allow
for much needed code clean-up.