LCM: IMPORTANT CHANGE FOR BUILDING TRILINOS/ALBANY/LCM ON FEDORA.
The convinience build scripts have been modified to use module files
instead of passing arguments with the build type, tool chain, etc.

This is so that the environment persists after a build and therefore
paths to libraries and executables are consistent with the latest
build.

CHANGES:

In ~/.bashrc, the following variables are needed now:

export LCM_DIR=~/LCM
export MODULEPATH=$LCM_DIR/Albany/doc/LCM/modulefiles

The LCM_DIR variable can point to other locations as well.

To see the available modules that correspond to different thread models,
compilers and builds:

module avail

which results in something like:

------------------ /home/amota/LCM/Albany/doc/LCM/modulefiles ------------------
openmp-gcc-debug     pthreads-gcc-release serial-gcc-debug
openmp-gcc-profile   pthreads-gcc-small   serial-gcc-profile
openmp-gcc-release   serial-clang-debug   serial-gcc-release
openmp-gcc-small     serial-clang-profile serial-gcc-small
pthreads-gcc-debug   serial-clang-release
pthreads-gcc-profile serial-clang-small

Then to set the environment for the appropriate build, for instance:

module load serial-gcc-release

And the first time:

cd ~/LCM
./clean-config-build.sh trilinos 72
./clean-config-build.sh albany 72

So that the number of parameters is now two instead of five as before.

After the first build, as usual:

./build.sh trilinos 72
./build.sh albany 72

All the rest remains the same.