GOAL: fix dashboard
ComputeHierarchicBCs used to have a check to
make sure that the problem was a "GOAL" problem.
If it wasn't, the function would return and
do nothing. For some reason, I deleted this check in
my last commit. This breaks pretty much every
other example in Albany when ENABLE_GOAL is turned
on by trying to find a GOAL-specific parameterlist
entry that does not exist, in general.
This commit should rectify the situation.