GOAL: better coordinate handling in disc
Previously the GOAL discretization was temporarily
changing the coordinate field of the apf mesh to
be the higher order solution field while calls
to updateMesh were called, then switching it
back to the appropriate linear field.

This hack has been removed so that the coordinate
array is now actually only associated with the
vertices of elements, rather than the nodes
of elements.