Little mod to EvaluatorUtils and GatherScalarNodalParameter
Added an optional parameter to the 'constructGatherScalarNodalParameter' method. The optional parameter is the field name. This is in analogy with what's done in StateManager, where the state and field names may differ. The constructor in GatherScalarNodalParameter has also been modified accordingly, so that it checks whether or not the input parameter list contain a 'Field Name' parameter.