Mod in STK meshes
The side set meshes may not be extracted as SideSetSTKMeshStruct, but may
also be loaded from file. This helps in case of a restart (all fields are
loaded correctly from the proper mesh). Also, for SideSetSTK meshes, we can
now also pass the adaptation parameters (they were not passed in the past
implementation).