Refactoring of FELIX::ResponseSurfaceVelocityMismatch
The evaluator now makes use of side set fields. Moreover, since the two sideset may have different layouts
we had to add the possibility of creating the response passing multiple versions of Albany::Layouts. To
this end, we had to add mulitiple layouts capabilities to Albany::ResponseUtilities. Obviously, for backward
compatibility, one can still pass a single Albany::Layout to ResponseUtilities. However, if needed, one can
pass a map string->layout with all the needed Albany::Layouts. In ResponseSurfaceVelocityMismatch, the key
of this map is the side set name.