Minor mod in (some of) the SideSet evaluators
When building the map sideNode->cellNode, we need to consider that the
number of nodes on each side of the cell may differ. This does not happen
on Tetrahedra or Hexahedra, but it does on Wedges.

However, notice that each sideset should contain only one shape, so during
the evaluation of the fields, the processed sides will always have the
same number of nodes!