Mod to StokesFO (this commit will be reverted!)
The basal friction field can now be loaded as a 3D field (taken from
the extruded field in the extruded mesh) and then projected back in 2D

This looks like a waste of time, and it is! We added this just to make some
checks. Don't use it on a real problem. Just load the 2D field...