LCM: Adding script to calculate the condition number mm matrices
This is a crude beginning. We just wanted to use scipy instead of Matlab
so that we can have more portability in testing the condition number.

So - the log10(condition number) for the beginning of MultiSlipHard with only the active sets is on order 5.
We should be to invert that. That said, the solver does nothing for 12 iterations.
That's not good. We need to examine both the preconditioner and Belos.