Adding a new option to Albany, specified by the Scaling parameter list:
<ParameterList name="Scaling">
    <Parameter name="Scale" type="double" value="1e9"/>
  </ParameterList>

"Scale" is a double value by which the non-Dirichlet entries of the global
Jacobian and Residual will be scaled.  The default value is 1.0.

For some problems, in particular, LCM, there are often solver issues due to the scales
associated and the way Dirichlet BCs are implemented in Albany (1s are put on the main
diagonal of the Jacobian for the Dirichlet dofs).  If the Jacobian entries corresponding
to free dofs are large, the Jacobian is very badly conditioned due to the 1s on the main
diagonal at the Dirichlet dofs.  By selecting a scale that is on the order of
the values of the Jacobian, the condition number can be improved greatly.

For now, the scale is set a priori and constant for all LOCA/Newton steps.