Export template signatures in ConstitutiveModelInterface_Def.hpp
Nightly tests using NVCC on shannon occasionally time out compiling ConstitutiveModelInterface_Def.hpp.
This file is a factory pattern, and includes headers (containing template
expressions) of many of the models in LCM/models. Use of the C++11 "export"
template keyword tells the compiler that the templates are explicitly instantiated
elsewhere (in this case, in the model *.cpp files), perhaps eliminating work
spent on template processing within the factory pattern.

This commit adds the export keyword to models used in the factory file to allow further testing of
compile behavior on shannon.