Temporarily mitigate long compile time using NVCC.
Albany can take several hours to compile using NVCC 7.5.7.
This results in the nightly build timing out on shannon. I
tracked this down to the functor body for NeohookeanKernel in the file
ParallelNeohookeanModel_Def.hpp . For now, I added a temporary
build flag:

-DNIGHTLY_TESTING:BOOL=ON

that turns off the code in the body of the functor. Enabling this option
decreases overall compile time to 15 minutes on shannon, but the ParallelNeohookean
model obviously will not work properly with this enabled.

TODO: Work with Alejandro to re-enable the desired functionality with reduced compile
times.

Note: This commit "backs out" changes made in

1318c87be77c593c9b71a08a35bda4af3ae4dad8

as this did not end up being an issue that could be addressed using ETI within Albany.