Added new feature to mesh/discretization
We added the map side3D_node_lid->cell2D_node_lid and refactored the map side3D_id->cell2D_id. The
reason for the new map is that the local id of a node within a 3D side may be different from the
local id of the same node within the 2D cell. In other words, the 2D cell and the 3D side may
(and in general do) have different local numerations. Therefore, when transferring data from 3D to 2D
or viceversa, we need to factor this in, otherwise we end up with data shuffled within the side/cell.