Mods to felix GIS input files.
Now the tests pass. The numbers obtained are exactly the same if the flow
rate type is Uniform, but differ if it is temperature based. This is because
the temperature is now interpolated from nodes to cell. This means that the
result is an average over the tetrahedron, while before the cell value was
computed averaging the nodal values over the prism during the extrusion.