Progress on proper tangent for Albany-Peridigm optimization-based coupling.
Hooked up plumbing to copy the tangent stiffness matrix from Peridigm to Albany.
The big catch is that the graph is not currently created properly, and instead
storage for all possible entries in rows corresponding to peridynamic nodes are
allocated.  This results in a ton of useless zeros in the matrix and is a horrible
idea, but getting it right will take some work.  It's not wrong, just slow and eats
a lot of memory.

For now, tested only in serial for pure peridynamics problems.  Outperforms
Matrix-Free and Finite Difference options and works for both Albany and AlbanyT.