Refactor of computeGraphs() for STK meshes to allow for insertion of nonzeros related to peridynamics for Albany-Peridigm coupling.
This version of the code creates a mathematically correct tangent stiffness matrix, but has
way too many nonzeros allocated in rows corresponding to peridynamic nodes.  The functions
that create the graphs for the stiffness matrix have been refactored to facilitate the next
step, which it to allocate the proper nonzero entries for peridynamic nodes.

 LocalWords:  src Untracked cmake albany linux